requirements.txt
amqp==5.2.0
#asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
#Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-paypal==2.1
djangorestframework-simplejwt==5.3.1
idna==3.7
kombu==5.4.0
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
six==1.16.0
#sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13


